enhancement: modular widgets in notifications
