+++ lib/features/notifications/widgets/detail_row.dart
@@ -1,0 +1,51 @@
+import 'package:flutter/material.dart';
+import 'package:heroicons/heroicons.dart';
+import 'package:mostro_mobile/core/app_theme.dart';
+
+class DetailRow extends StatelessWidget {
+  final String label;
+  final String value;
+  final HeroIcons icon;
+
+  const DetailRow({
+    super.key,
+    required this.label,
+    required this.value,
+    required this.icon,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.symmetric(vertical: 2),
+      child: Row(
+        children: [
+          HeroIcon(
+            icon,
+            style: HeroIconStyle.outline,
+            size: 14,
+            color: AppTheme.textSecondary.withValues(alpha: 0.7),
+          ),
+          const SizedBox(width: 8),
+          Text(
+            '$label:',
+            style: Theme.of(context).textTheme.bodySmall?.copyWith(
+              fontWeight: FontWeight.w500,
+              color: AppTheme.textSecondary,
+            ),
+          ),
+          const SizedBox(width: 4),
+          Expanded(
+            child: Text(
+              value,
+              style: Theme.of(context).textTheme.bodySmall?.copyWith(
+                fontFamily: value.contains('npub') || value.contains('#') ? 'monospace' : null,
+                color: AppTheme.textPrimary,
+              ),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
